[rename] Modify a directory structure
Github issue #34
