+++ src/App.tsx
@@ -3,7 +3,7 @@ import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
 import MainPage from './page/MainPage';
 import MyPage from './page/Mypage';
 import ConvertPage from './page/ConvertPage';
-import NavBar from './component/Navbar';
+import NavBar from './component/base/Navbar';
 
 const App = () => {
   return (
