+++ src/component/convert/novel/CharacterChip.tsx
