+++ src/component/mainpage/MainIndicator.tsx
